+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_form_training_client.py
@@ -66,7 +66,7 @@ class FormTrainingClient(object):
         )
 
     @distributed_trace
-    def begin_train_model(self, training_files_url, use_labels=False, **kwargs):
+    def begin_train_model(self, training_files_url, use_training_labels=False, **kwargs):
         # type: (str, Optional[bool], Any) -> LROPoller
         """Create and train a custom model. The request must include a `training_files_url` parameter that is an
         externally accessible Azure storage blob container Uri (preferably a Shared Access Signature Uri).
@@ -74,7 +74,7 @@ class FormTrainingClient(object):
         'image/jpeg', 'image/png', 'image/tiff'. Other type of content in the container is ignored.
 
         :param str training_files_url: An Azure Storage blob container's SAS URI.
-        :param bool use_labels: Whether to train with labels or not. Corresponding labeled files must
+        :param bool use_training_labels: Whether to train with labels or not. Corresponding labeled files must
             exist in the blob container.
         :keyword str prefix: A case-sensitive prefix string to filter documents for training.
             Use `prefix` to filter documents themselves, or to restrict sub folders for training
@@ -104,7 +104,7 @@ class FormTrainingClient(object):
         response = self._client.train_custom_model_async(  # type: ignore
             train_request=TrainRequest(
                 source=training_files_url,
-                use_label_file=use_labels,
+                use_label_file=use_training_labels,
                 source_filter=TrainSourceFilter(
                     prefix=kwargs.pop("prefix", ""),
                     include_sub_folders=kwargs.pop("include_sub_folders", False),