+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/async_samples/sample_train_model_without_labels_async.py
@@ -42,7 +42,7 @@ class TrainModelWithoutLabelsSampleAsync(object):
             self.endpoint, AzureKeyCredential(self.key)
         ) as form_training_client:
 
-            # Default for train_model is `use_labels=False`
+            # Default for train_model is `use_training_labels=False`
             model = await form_training_client.train_model(self.container_sas_url)
 
             # Custom model information