+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -5,6 +5,7 @@
 **Breaking Changes**
 
 - `training_files` parameter of `begin_train_model` is renamed to `training_files_url`
+- `use_labels` parameter of `begin_train_model` is renamed to `use_training_labels`
 - `list_model_infos` method has been renamed to `list_custom_models`
 
 ## 1.0.0b2 (2020-05-06)